GraphOTSim/requirements.txt
ConfigArgParse==0.14.0
cvxopt==1.2.3
dgl-cu101==0.4.3.post2
jupyter==1.0.0
jupyter==1.0.0
notebook==5.7.8
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
#jupyter-tensorboard==0.1.8
jupyter_contrib_nbextensions
lap==0.4.0
networkx==2.4
notebook==5.7.8
opencv-python==4.4.0.42
Pillow==7.2.0
scipy==1.1.0
seaborn==0.9.0
torch==1.4.0
torchvision==0.5.0
torch-geometric==1.6.1
torch-scatter==2.0.4
torch-sparse==0.6.1
tqdm==4.48.2
texttable
